requirements.txt
numpy==1.21.5
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
radiant_mlhub==0.4.1
setuptools==61.2.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
wandb
pytest